Enhanced NNUE model loading and file checking
- Added path normalization and NNUE file header checking in evaluate.cpp to improve model loading reliability.
- Updated the init_nnue function to add file size and version checks and provide detailed debugging information during loading.
- Declared new path normalization and file header checking functions in evaluate.h.
- Optimized type conversion in nnue/evaluate_nnue.cpp to ensure accurate data processing.

These changes improve the traceability and user experience of the NNUE model loading process.

+++ src/nnue/nnue_common.h
@@ -161,11 +161,11 @@ namespace Stockfish::Eval::NNUE {
           {
             for (; i + 1 < sizeof(IntType); ++i)
             {
-                u[i] = v;
+                u[i] = static_cast<std::uint8_t>(v);
                 v >>= 8;
             }
           }
-          u[i] = v;
+          u[i] = static_cast<std::uint8_t>(v);
 
           stream.write(reinterpret_cast<char*>(u), sizeof(IntType));
       }
